MEAO update:
Key Changes:
maxChange Parameter:
Added "maxChange": 1 to the API request parameters to limit the number of changes to 1.

Filter Journeys:
The function now selects the journey with the fewest changes (in this case, 1 or fewer).

Which means ...
Performance Improvement:
By limiting the number of changes, the TFL API will return fewer journey options, which reduces the time spent processing and filtering results.

This should make the app faster, especially when calculating routes for multiple users.